Fixed posix_thread issues
Issue with core affinity, priority and the creation of threads. Setting inherit scheduler to PTHREAD_EXPLICIT_SCHED allows created thread to take on the attributes given to it rather than those of the thread creating it.

Also, note on thread priority:
If there is an issue with the test on thread priority it may mean that there is a system limit on the priority of real time scheduling policies (such as SCHED_RR and SCHED_FIFO). Posix_thread uses SCHED_RR for now. You can check the limit of priority in the shell with command 'ulimit -r'. To change the limit, open /etc/security/limits.conf and insert '* - rtprio unlimited' at the end of the file.