Fixed non thread-safe unlock() of posix_mutex (as a binary semaphore)
Previously there was the threat of a possibly race condition in posix_mutex::unlock().
Have an extra "mutex" (binary semaphore), unlockHandle_, which serves as a gate in the unlock() function.
When handle_ is unlockable, unlockHandle_ is lockable. They have opposite values.
lock(unlockHandle_) makes sure we execute sem_getvalue() and sem_post() in unlock() atomically.
Also made the ctor thread-safe with the constructing of handle_ and unlockHandle_.